Déplace bouton Outils dans le header principal, à côté de Paramètres
- Supprime Outils de CompetitionNav (bouton + modals + état)
- Ajoute Outils dans App.tsx header, visible quand une compétition est active
- CompetitionView remonte remoteServerUrl/arenaCount via onRemoteServerChange
